Move the benchmarks into separate library
