+++ src/benchmark/util.cc
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "pymatching/perf/util.perf.h"
+#include "benchmark/util.h"
